Add files via upload
The Adventure works Cycle is a manufacturing company and it manufactures the cycles , composite Bicycles , Mountain -100 black, Mountain -100 silver  Which contributes Top sales and it sells to other countries like North America, European and Asia.
In 2000s, Adventure Works Cycles bought a small manufacturing plant in Mexico. Which manufactures several critical subcomponents for the Adventure Works Cycles product line. These subcomponents are shipped to the Bothell location for final product assembly. In 2001, this manufacturing plant became the sole manufacturer and distributor of the touring bicycle product group.